Move database related code to utilities.database.
